Add attribute.ts
